style(split): db get position
