+++ backend/matcha/db/db.py
@@ -23,36 +23,6 @@ def db_get_id_password_where_username(username):
     return db_fetchone("SELECT id,password FROM users WHERE username = %s", (username,))
 
 
-def db_get_position(id_user):
-    return db_fetchone("SELECT latitude,longitude FROM users WHERE id = %s", (id_user,))
-
-
-def db_update_position(id_user: int, latitude: float, longitude: float):
-    error_msg = db_query(
-        "UPDATE users SET (latitude, longitude) = (%s, %s) WHERE id = %s",
-        (latitude, longitude, id_user),
-    )
-
-    if error_msg:
-        return error_msg
-
-
-def db_update_address(id_user: int, address: dict):
-    error_msg = db_query(
-        "UPDATE users SET (street_name, street_number, city, country) = (%s, %s, %s, %s) WHERE id = %s",
-        (
-            address["road"],
-            address["house_number"],
-            address["town"],
-            address["country"],
-            id_user,
-        ),
-    )
-
-    if error_msg:
-        return error_msg
-
-
 def db_set_user_email(id_user, email):
     error_msg = db_query(
         "UPDATE users SET email = %s where id = %s",
